+++ regionmask/__init__.py
@@ -1,16 +1,20 @@
 # flake8: noqa
 
-import os
-
-import pkg_resources
-
 from . import core, defined_regions
 from .core._geopandas import from_geopandas, mask_3D_geopandas, mask_geopandas
 from .core.plot import plot_3D_mask
 from .core.regions import Regions, _OneRegion
 
 try:
-    __version__ = pkg_resources.get_distribution("regionmask").version
+    from importlib.metadata import version as _get_version
+except ImportError:
+    # importlib.metadata not available in python 3.7
+    import pkg_resources
+
+    _get_version = lambda pkg: pkg_resources.get_distribution(pkg).version
+
+try:
+    __version__ = _get_version("regionmask")
 except Exception:
     # Local copy or not installed with setuptools.
     # Disable minimum version checks on downstream libraries.